// src/types/allegro.ts
export interface AllegroRefund {
  id: string;
  payment: {
    id: string;
  };
  reason: {
    id: string;
    name: string;
  };
  status: 'PENDING' | 'SUCCEEDED' | 'FAILED';
  value: {
    amount: string;
    currency: 'PLN';
  };
  lineItems: {
    lineItemId: string;
    quantity: number;
    amount: {
      amount: string;
      currency: 'PLN';
    };
  }[];
  createdAt: string;
  updatedAt: string;
}